Added overloaded constructor and mode types
Added two "constants" these are public static final variables and make
it easier to read code that uses cryptic numbers like 1 to denote a
mode. I've added a ALIEN_MODE and a TOMMY_MODE (although this might
point towards the need to create a subclass of ship...)
Overloaded the constructor so that a ship can be constructed without
specifying a mode. This non-mode constructor "chains" to the original
constructor, passing in ALIEN_MODE