A lot of things
Added a powerup class and sprite 1 of the class: ammo. I commented out
the missile's collision for ship because it was not working. I also made
my game a lot better. Finally, i added several didHit functions that can
probably be a lot cleaner for the ship and missile class.